+++ example/lib/src/model/test_model.dart
@@ -10,7 +10,4 @@ class TestMode {
   }
 }
 
-class TestMode2
-{
-
-}
+class TestMode2 {}