+++ example1/lib/src/model/test_model.dart
@@ -19,7 +19,4 @@ class TestMode {
   }
 }
 
-class TestMode2
-{
-
-}
+class TestMode2 {}