+++ example1/lib/src/model/test_model1.dart
@@ -18,8 +18,4 @@ class TestMode1 {
   }
 }
 
-
-class TestMode3
-{
-
-}
+class TestMode3 {}